Stored_Functions.sql push. File didn't push. Re-upload. Uses the phpProject database.
Updated login check - Made it check to see if the username exists.
Admin_Profile_Change - Allows for resetting passwords, salts, and usernames as well as the functionality of Profile_Change.
Profile_Change - Allows changes to emails and passwords.
Create_Users - Allows creation of user, salt related to user profile must be passed in.